Refactor, mostly sorting into packages
