+++ src/main/java/kniaz/logic/handler/CommandHandler.java
@@ -1,12 +1,11 @@
-package main.logic.handler;
+package kniaz.logic.handler;
 
 import java.util.List;
 import java.util.Map;
 
 import exceptions.syntax.MissingUnnamedArgsException;
 import exceptions.syntax.UnknownCommandException;
-import main.KniazSession;
-
+import kniaz.KniazSession;
 
 
 /**
