+++ src/main/java/kniaz/logic/handler/ListHandler.java
@@ -1,10 +1,9 @@
-package main.logic.handler;
+package kniaz.logic.handler;
 
 import java.util.List;
 import java.util.Map;
 
-import main.KniazSession;
-
+import kniaz.KniazSession;
 
 
 /**
