+++ src/main/java/kniaz/logic/handler/ToDoHandler.java
@@ -1,10 +1,10 @@
-package main.logic.handler;
+package kniaz.logic.handler;
 
 import java.util.List;
 import java.util.Map;
 
 import exceptions.syntax.MissingUnnamedArgsException;
-import main.KniazSession;
+import kniaz.KniazSession;
 import task.Task;
 import task.ToDo;
