+++ src/main/java/ui/gui/MainWindow.java
@@ -1,3 +1,5 @@
+package ui.gui;
+
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -6,12 +8,14 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import kniaz.Kniaz;
 import ui.UiCommand;
+
 //@@author Jeffry Lium
 //Reused from https://se-education.org/guides/tutorials/javaFxPart4.html
 // with minor modifications
 /**
- * Controller for MainWindow. Provides the layout for the other controls.
+ * Controller for ui.gui.MainWindow. Provides the layout for the other controls.
  */
 public class MainWindow extends AnchorPane {
     @FXML
