+++ src/main/resources/view/MainWindow.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0"
             prefWidth="400.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="MainWindow">
+            fx:controller="ui.gui.MainWindow">
     <children>
         <TextField fx:id="userInput" layoutY="558.0"
                    onAction="#grabUserInput" prefHeight="41.0"
